+++ src/oidc_provider/error.clj
@@ -17,6 +17,8 @@
       (derive ::invalid-grant ::request-error)
       (derive ::invalid-client-metadata ::request-error)
       (derive ::unsupported-grant-type ::request-error)
+      (derive ::unsupported-response-type ::request-error)
+      (derive ::invalid-scope ::request-error)
       (derive ::invalid-client ::auth-error)
       (derive ::invalid-token ::auth-error)))